Increase UnitInfo speed and group_speed limit to 65535.

+++ src/game_manager.cpp
@@ -2884,7 +2884,7 @@ void GameManager_ProcessCheatCodes() {
                     if (new_unit_values->GetAttribute(ATTRIB_SPEED) <
                         2 * selected_unit_values->GetAttribute(ATTRIB_SPEED)) {
                         new_unit_values->SetAttribute(
-                            ATTRIB_SPEED, std::min(2 * selected_unit_values->GetAttribute(ATTRIB_SPEED), UINT8_MAX));
+                            ATTRIB_SPEED, std::min(2 * selected_unit_values->GetAttribute(ATTRIB_SPEED), UINT16_MAX));
                     }
 
                     if (new_unit_values->GetAttribute(ATTRIB_ROUNDS) <
